feat: move form template routes from public to admin
