+++ src/app/routes/api/v3/admin/forms/admin-forms.routes.ts
@@ -12,6 +12,7 @@ import { AdminFormsPresignRouter } from './admin-forms.presign.routes'
 import { AdminFormsPreviewRouter } from './admin-forms.preview.routes'
 import { AdminFormsSettingsRouter } from './admin-forms.settings.routes'
 import { AdminFormsSubmissionsRouter } from './admin-forms.submissions.routes'
+import { AdminFormTemplatesRouter } from './admin-forms.templates.routes'
 import { AdminFormsTwilioRouter } from './admin-forms.twilio.routes'
 
 export const AdminFormsRouter = Router()
@@ -30,3 +31,4 @@ AdminFormsRouter.use(AdminFormsPreviewRouter)
 AdminFormsRouter.use(AdminFormsPresignRouter)
 AdminFormsRouter.use(AdminFormsLogicRouter)
 AdminFormsRouter.use(AdminFormsTwilioRouter)
+AdminFormsRouter.use(AdminFormTemplatesRouter)
