+++ src/app/routes/api/v3/forms/public-forms.routes.ts
@@ -4,7 +4,6 @@ import { PublicFormsAuthRouter } from './public-forms.auth.routes'
 import { PublicFormsFeedbackRouter } from './public-forms.feedback.routes'
 import { PublicFormsFormRouter } from './public-forms.form.routes'
 import { PublicFormsSubmissionsRouter } from './public-forms.submissions.routes'
-import { PublicFormTemplatesRouter } from './public-forms.templates.routes'
 import { PublicFormsVerificationRouter } from './public-forms.verification.routes'
 
 export const PublicFormsRouter = Router()
@@ -14,4 +13,3 @@ PublicFormsRouter.use(PublicFormsFeedbackRouter)
 PublicFormsRouter.use(PublicFormsFormRouter)
 PublicFormsRouter.use(PublicFormsAuthRouter)
 PublicFormsRouter.use(PublicFormsVerificationRouter)
-PublicFormsRouter.use(PublicFormTemplatesRouter)
